@@ -1,4 +1,7 @@
-﻿using System.Windows;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Windows;
 
 namespace TodoListClient
 {
